API/requirements.txt
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.0
Flask==2.2.2
Flask-Cors==3.0.10
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
numpy==1.24.1
pandas==1.5.3
protobuf==3.20.3
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
Werkzeug==2.2.2
zipp==3.12.0
